requirements.txt
aiohttp
mnemonic
pynacl
cryptography
aiofiles
msgpack
